==> module_03/1-GetData.ipynb <==
# We use names-dataset python package to get country-wise popular names. The package uses the massive Facebook dump of 533M users.

from names_dataset import NameDataset
import tqdm, os
import pandas as pd

DATA_DIR = "data/"

nd = NameDataset()

countries = nd.get_country_codes(alpha_2=False)

Fetching lastnames and gender-wise firstnames for each country

data = {}
with tqdm.tqdm(total=len(countries)) as pbar:
    for country in countries:
        data[country.name] = []
        temp = nd.get_top_names(
            n=100000, use_first_names=True, gender="M", country_alpha2=country.alpha_2
        )
        data[country.name].append(temp[country.alpha_2]["M"])
        temp = nd.get_top_names(
            n=100000, use_first_names=True, gender="F", country_alpha2=country.alpha_2
        )
        data[country.name].append(temp[country.alpha_2]["F"])
        temp = nd.get_top_names(
            n=100000, use_first_names=False, country_alpha2=country.alpha_2
        )
        data[country.name].append(temp[country.alpha_2])
        pbar.update(1)

Let us focus on the following countries only

country_names = {
    "India": "india",
    "United States": "usa",
    "United Kingdom": "uk",
    "Nigeria": "nigeria",
    "United Arab Emirates": "uae",
    "Morocco": "morocco",
    "Netherlands": "netherlands",
    "Peru": "peru",
    "Germany": "germany",
    "France": "france",
    "Italy": "italy",
    "Bangladesh": "bangladesh",
    "Singapore": "singapore",
    "Hong Kong": "hongkong",
    "South Africa": "southafrica",
    "Spain": "spain",
    "Mexico": "mexico",
    "Malaysia": "malaysia",
    "Canada": "canada",
    "Indonesia": "indonesia",
}

We only want to use the names which contain only the english characters: a-z

dataset = []
for country in country_names:
    for name in data[country][0]:
        if name.isascii() and name.isalpha() and len(name) > 0:
            dataset.append([country_names[country], "firstname", "male", name.lower()])
    for name in data[country][1]:
        if name.isascii() and name.isalpha() and len(name) > 0:
            dataset.append(
                [country_names[country], "firstname", "female", name.lower()]
            )
    for name in data[country][0]:
        if name.isascii() and name.isalpha() and len(name) > 0:
            dataset.append(
                [country_names[country], "lastname", "neutral", name.lower()]
            )

df = pd.DataFrame(dataset, columns=["country", "nametype", "gender", "name"])
df.to_csv(DATA_DIR + "dataset.csv", index=False)
df
