==> module_03/2-PreparePre-TrainingDataset.ipynb <==
# Prepares the dataset for pre-training. This includes cleaning of dataset,  creating vocabulary, defining encoder and decoder functions.

import os, time, math, pickle
import numpy as np, pandas as pd

DATA_DIR = "data/"
seed = 999

Cleaning the data

df = pd.read_csv(DATA_DIR + "dataset.csv")
df.dropna(inplace=True)
df["length"] = df["name"].apply(lambda x: len(x))
df = df[(df["nametype"] == "firstname") & (df["length"] < 15) & (df["length"] > 2)]
names = df["name"].tolist()
print("length of dataset in characters =", sum([len(x) for x in names]))

Vocabulary and encoding/decoding

chars = set()
for name in names:
    chars.update(list(name))
chars = sorted(list(chars))
chars += ["*", "{", "}", "=", "S", "E", "G", "M", "F", "C", "I", "O"]
vocab_size = len(chars)
print("all the unique characters:", "".join(chars))
print(f"vocab size: {vocab_size:,}")

stoi = {ch: i for i, ch in enumerate(chars)}
itos = {i: ch for i, ch in enumerate(chars)}


def encode(s):
    return [stoi[c] for c in s]


def decode(l):
    return "".join([itos[i] for i in l])

splitting data into training and validation

train_df = df.sample(frac=0.8, random_state=seed)
val_df = df.drop(train_df.index)
val_df = val_df.sample(frac=1)
train_df.to_csv(DATA_DIR + "train.bin", index=False)
val_df.to_csv(DATA_DIR + "val.bin", index=False)
print("training names:", len(train_df), "validation names:", len(val_df))

save metadata

meta = {
    "vocab_size": vocab_size,
    "itos": itos,
    "stoi": stoi,
}
with open(DATA_DIR + "meta.pkl", "wb") as f:
    pickle.dump(meta, f)

itos

stoi
