==> module_03/5-Embeddings.ipynb <==
import os, pickle, torch
from model_exercise1_solution import GPT
from matplotlib import pyplot as plt

DATA_DIR = "data/"
MODEL_DIR = "best_models/"
CHECKPOINT = "base.pt"
device = "mps"
device = "cuda" if torch.cuda.is_available() else "cpu"
print("device =", device)

checkpoint = torch.load(MODEL_DIR + CHECKPOINT, map_location=device)
config = checkpoint["config"]
model = GPT(config)
state_dict = checkpoint["model"]
unwanted_prefix = "_orig_mod."
for k, v in list(state_dict.items()):
    if k.startswith(unwanted_prefix):
        state_dict[k[len(unwanted_prefix) :]] = state_dict.pop(k)
model.load_state_dict(state_dict)
print("best val loss =", checkpoint["best_val_loss"].item())

with open(DATA_DIR + "meta.pkl", "rb") as f:
    meta = pickle.load(f)
stoi, itos = meta["stoi"], meta["itos"]
encode = lambda s: [stoi[c] for c in s]
decode = lambda l: "".join([itos[i] for i in l])

dims = torch.randint(config["n_embd"], (1, 2))
dim1, dim2 = dims[0, 0].item(), dims[0, 1].item()
for c_i in range(26):
    x, y = (
        model.transformer.wte.weight[c_i, dim1].item(),
        model.transformer.wte.weight[c_i, dim2].item(),
    )
    plt.scatter(x, y, marker="x", color="red", s=4)
    plt.text(x + 0.001, y + 0.001, itos[c_i], fontsize=7)
plt.xlabel("dimension: " + str(dim1))
plt.ylabel("dimension: " + str(dim2))
plt.show()

dims = torch.randint(config["n_embd"], (1, 2))
dim1, dim2 = dims[0, 0].item(), dims[0, 1].item()
for c_i in range(26):
    x, y = (
        model.transformer.wte.weight[c_i, dim1].item(),
        model.transformer.wte.weight[c_i, dim2].item(),
    )
    plt.scatter(x, y, marker="x", color="red", s=4)
    plt.text(x + 0.001, y + 0.001, itos[c_i], fontsize=7)
plt.xlabel("dimension: " + str(dim1))
plt.ylabel("dimension: " + str(dim2))
plt.show()

dims = torch.randint(config["n_embd"], (1, 2))
dim1, dim2 = dims[0, 0].item(), dims[0, 1].item()
for c_i in range(26):
    x, y = (
        model.transformer.wte.weight[c_i, dim1].item(),
        model.transformer.wte.weight[c_i, dim2].item(),
    )
    plt.scatter(x, y, marker="x", color="red", s=4)
    plt.text(x + 0.001, y + 0.001, itos[c_i], fontsize=7)
plt.xlabel("dimension: " + str(dim1))
plt.ylabel("dimension: " + str(dim2))
plt.show()

dims = torch.randint(config["n_embd"], (1, 2))
dim1, dim2 = dims[0, 0].item(), dims[0, 1].item()
for c_i in range(26):
    x, y = (
        model.transformer.wte.weight[c_i, dim1].item(),
        model.transformer.wte.weight[c_i, dim2].item(),
    )
    plt.scatter(x, y, marker="x", color="red", s=4)
    plt.text(x + 0.001, y + 0.001, itos[c_i], fontsize=7)
plt.xlabel("dimension: " + str(dim1))
plt.ylabel("dimension: " + str(dim2))
plt.show()

for c_i in range(26):
    sim = {}
    for c_j in range(26):
        if c_j == c_i:
            continue
        sim[c_j] = round(
            torch.dot(
                model.transformer.wte.weight[c_i], model.transformer.wte.weight[c_j]
            ).item(),
            2,
        )
    sim = sorted(sim.items(), key=lambda item: item[1], reverse=True)
    print(
        "similar chars to "
        + itos[c_i]
        + ": "
        + itos[sim[0][0]]
        + "("
        + str(sim[0][1])
        + ")"
        + ", "
        + itos[sim[1][0]]
        + "("
        + str(sim[1][1])
        + ")"
        + ", "
        + itos[sim[2][0]]
        + "("
        + str(sim[2][1])
        + ")"
    )

stoi["{"], stoi["}"]

for c_i in [27, 28]:
    sim = {}
    for c_j in range(26):
        if c_j == c_i:
            continue
        sim[c_j] = round(
            torch.dot(
                model.transformer.wte.weight[c_i], model.transformer.wte.weight[c_j]
            ).item(),
            2,
        )
    sim = sorted(sim.items(), key=lambda item: item[1], reverse=True)
    print(
        "similar chars to "
        + itos[c_i]
        + ": "
        + itos[sim[0][0]]
        + "("
        + str(sim[0][1])
        + ")"
        + ", "
        + itos[sim[1][0]]
        + "("
        + str(sim[1][1])
        + ")"
        + ", "
        + itos[sim[2][0]]
        + "("
        + str(sim[2][1])
        + ")"
    )
