==> module_03/6-PrepareFine-TuningDatasets.ipynb <==
import os, time, math, pickle, random
import numpy as np, pandas as pd

DATA_DIR = "data/"

train_df = pd.read_csv(DATA_DIR + "train.bin")
train_df[train_df["name"].str.startswith("a")].sample(10000)[["name"]].to_csv(
    DATA_DIR + "startswith_a_train.bin", index=False
)
val_df = pd.read_csv(DATA_DIR + "val.bin")
val_df[val_df["name"].str.startswith("a")].sample(2000)[["name"]].to_csv(
    DATA_DIR + "startswith_a_val.bin", index=False
)

train_df = pd.read_csv(DATA_DIR + "train.bin")
train_df[train_df["name"].str.endswith("a")].sample(10000)[["name"]].to_csv(
    DATA_DIR + "endswith_a_train.bin", index=False
)
val_df = pd.read_csv(DATA_DIR + "val.bin")
val_df[val_df["name"].str.endswith("a")].sample(2000)[["name"]].to_csv(
    DATA_DIR + "endswith_a_val.bin", index=False
)

train_df = pd.read_csv(DATA_DIR + "train.bin")
indian = train_df[train_df["country"] == "india"].sample(3000)
non_indian = train_df[train_df["country"] != "india"].sample(10000)
train_df = pd.concat([indian, non_indian])
train_df.sample(frac=1)[["name", "country"]].to_csv(
    DATA_DIR + "indian_classification_train.bin", index=False
)
val_df = pd.read_csv(DATA_DIR + "val.bin")
indian = val_df[val_df["country"] == "india"].sample(500)
non_indian = val_df[val_df["country"] != "india"].sample(1500)
val_df = pd.concat([indian, non_indian])
val_df.sample(frac=1)[["name", "country"]].to_csv(
    DATA_DIR + "indian_classification_val.bin", index=False
)

train_df = pd.read_csv(DATA_DIR + "train.bin")
male = train_df[train_df["gender"] == "male"].sample(6000)
female = train_df[train_df["gender"] == "female"].sample(4000)
train_df = pd.concat([male, female])
train_df.sample(frac=1)[["name", "gender"]].to_csv(
    DATA_DIR + "gender_classification_train.bin", index=False
)
val_df = pd.read_csv(DATA_DIR + "val.bin")
male = val_df[val_df["gender"] == "male"].sample(1200)
female = val_df[val_df["gender"] == "female"].sample(800)
val_df = pd.concat([male, female])
val_df.sample(frac=1)[["name", "gender"]].to_csv(
    DATA_DIR + "gender_classification_val.bin", index=False
)

data = []
train_df = pd.read_csv(DATA_DIR + "train.bin")
indian = train_df[train_df["country"] == "india"]
indian_male = indian[indian["gender"] == "male"].sample(3000)
indian_female = indian[indian["gender"] == "female"].sample(2500)
non_indian = train_df[train_df["country"] != "india"]
non_indian_male = non_indian[non_indian["gender"] == "male"].sample(3000)
non_indian_female = non_indian[non_indian["gender"] == "female"].sample(2500)
train_df = pd.concat([indian_male, indian_female, non_indian_male, non_indian_female])
for index, row in train_df.iterrows():
    prefix_size = min(len(row["name"]), random.choice([1, 2]))
    prefix = row["name"][:prefix_size]
    suffix_size = min(len(row["name"]), random.choice([1, 2]))
    suffix = row["name"][-suffix_size:]
    country = "I" if row["country"] == "india" else "O"
    gender = "M" if row["gender"] == "male" else "F"
    data.append("S" + prefix + "0" + "{" + row["name"] + "}")
    data.append("E" + suffix + "0" + "{" + row["name"] + "}")
    data.append("G" + "{" + row["name"] + "0" + "=" + gender + "}")
    data.append("C" + "{" + row["name"] + "0" + "=" + country + "}")
train_df = pd.DataFrame(data, columns=["name"])
train_df = train_df.sample(frac=1)
train_df.to_csv(DATA_DIR + "instruction_tuning_train.bin", index=False)
data = []
val_df = pd.read_csv(DATA_DIR + "val.bin")
indian = val_df[val_df["country"] == "india"]
indian_male = indian[indian["gender"] == "male"].sample(600)
indian_female = indian[indian["gender"] == "female"].sample(400)
non_indian = val_df[val_df["country"] != "india"]
non_indian_male = non_indian[non_indian["gender"] == "male"].sample(600)
non_indian_female = non_indian[non_indian["gender"] == "female"].sample(400)
val_df = pd.concat([indian_male, indian_female, non_indian_male, non_indian_female])
for index, row in val_df.iterrows():
    prefix_size = min(len(row["name"]), random.choice([1, 2]))
    prefix = row["name"][:prefix_size]
    suffix_size = min(len(row["name"]), random.choice([1, 2]))
    suffix = row["name"][-suffix_size:]
    country = "I" if row["country"] == "india" else "O"
    gender = "M" if row["gender"] == "male" else "F"
    data.append("S" + prefix + "0" + "{" + row["name"] + "}")
    data.append("E" + suffix + "0" + "{" + row["name"] + "}")
    data.append("G" + "{" + row["name"] + "0" + "=" + gender + "}")
    data.append("C" + "{" + row["name"] + "0" + "=" + country + "}")
val_df = pd.DataFrame(data, columns=["name"])
val_df = val_df.sample(frac=1)
val_df.to_csv(DATA_DIR + "instruction_tuning_val.bin", index=False)

train_df = pd.read_csv(DATA_DIR + "train.bin").sample(10000)
train_df["name"] = train_df["name"].apply(lambda x: "{" + x + "0" + "=" + x[0] + "}")
train_df[["name"]].to_csv(DATA_DIR + "first_char_train.bin", index=False)
val_df = pd.read_csv(DATA_DIR + "val.bin").sample(2000)
val_df["name"] = val_df["name"].apply(lambda x: "{" + x + "0" + "=" + x[0] + "}")
val_df[["name"]].to_csv(DATA_DIR + "first_char_val.bin", index=False)

train_df = pd.read_csv(DATA_DIR + "train.bin").sample(10000)
train_df["name"] = train_df["name"].apply(lambda x: "{" + x + "0" + "=" + x[-1] + "}")
train_df[["name"]].to_csv(DATA_DIR + "last_char_train.bin", index=False)
val_df = pd.read_csv(DATA_DIR + "val.bin").sample(2000)
val_df["name"] = val_df["name"].apply(lambda x: "{" + x + "0" + "=" + x[-1] + "}")
val_df[["name"]].to_csv(DATA_DIR + "last_char_val.bin", index=False)
